Playing with from_json method to overcome JSON limitations
