+++ mlbstatsapi/models/people/people.py
@@ -1,25 +1,12 @@
-﻿from .person import Person
-from .attributes import PrimaryPosition
-from dataclasses import dataclass, field
-from typing import Union, Optional
+﻿from typing import List, Dict, Union, Any, Optional
+from .attributes import PitchHand, PrimaryPosition, BatSide, Status
+from mlbstatsapi import MlbObject
+from mlbstatsapi.models.stats import Stats
+from dataclasses import dataclass, field, InitVar
+
 
 @dataclass
-class Status:
-    """
-    A dataclass to hold player status
-
-    Attributes
-    ----------
-    code : str
-        code of the player
-    description : str
-        description of the status
-    """
-    code: str
-    description: str
-
-@dataclass 
-class Player:
+class Person:
     """
     A class to represent a Person.
 
@@ -27,8 +14,77 @@ class Player:
     ----------
     id : int
         id number of the person
-    full_name : str
+    full_name: str
         full_name of the person
+    """
+
+    id: int
+    link: str
+    mlb_class: str = "people"
+    primaryPosition: Union[PrimaryPosition, Dict[str, Any]] = field(default_factory=dict)
+    pitchHand: Union[PitchHand, Dict[str, Any]] = field(default_factory=dict)
+    batSide: Union[BatSide, Dict[str, Any]] = field(default_factory=dict)
+    stats: Union[Stats, list] = field(default_factory=list)
+    fullName: Optional[str] = None
+    firstName: Optional[str] = None
+    lastName: Optional[str] = None
+    primaryNumber: Optional[str] = None
+    birthDate: Optional[str] = None
+    currentTeam: Optional[str] = None
+    currentAge: Optional[str] = None
+    birthCity: Optional[str] = None
+    birthStateProvince: Optional[str] = None
+    height: Optional[str] = None
+    weight: Optional[int] = None
+    active: Optional[bool] = None
+    useName: Optional[str] = None
+    middleName: Optional[str] = None
+    boxscoreName: Optional[str] = None
+    nickName: Optional[str] = None
+    draftYear: Optional[int] = None
+    mlbDebutDate: Optional[str] = None
+    nameFirstLast: Optional[str] = None
+    nameSlug: Optional[str] = None
+    firstLastName: Optional[str] = None
+    lastFirstName: Optional[str] = None
+    lastInitName: Optional[str] = None
+    initLastName: Optional[str] = None
+    fullFMLName: Optional[str] = None
+    fullLFMName: Optional[str] = None
+    birthCountry: Optional[str] = None
+    pronunciation: Optional[str] = None
+    strikeZoneTop: Optional[float] = None
+    strikeZoneBottom: Optional[float] = None
+    nameTitle: Optional[str] = None
+    gender: Optional[str] = None
+    isPlayer: Optional[bool] = None
+    isVerified: Optional[bool] = None
+    nameMatrilineal: Optional[str] = None
+    deathDate: Optional[str] = None
+    deathCity: Optional[str] = None
+    deathCountry: Optional[str] = None
+
+    def __post_init__(self):
+        self.primaryPosition = PrimaryPosition(**self.primaryPosition) if self.primaryPosition else self.primaryPosition
+        self.pitchHand = PitchHand(**self.pitchHand) if self.pitchHand else self.pitchHand
+        self.batSide = BatSide(**self.batSide) if self.batSide else self.batSide
+        self.stats = [ Stats(**stat) for stat in self.stats ] if self.stats else self.stats
+
+
+
+
+
+@dataclass(kw_only=True)
+class Player(Person):
+    """
+    A class to represent a Player.
+
+    Attributes
+    ----------
+    jerseyNumber : str
+        id number of the person
+    position : PrimaryPosition
+        PrimaryPosition of the player
     status : 
         Status of the player
     parentTeamId : int
@@ -36,9 +92,54 @@ class Player:
     jerseyNumber: str
     parentTeamId: int
     position: Union[PrimaryPosition, dict] = field(default_factory=dict)
-    person: Union[Person, dict] = field(default_factory=dict)
     status: Union[Status, dict] = field(default_factory=dict)
+
     def __post_init__(self):
         self.position = PrimaryPosition(**self.position) if self.position else self.position
-        self.person = Person(**self.person) if self.person else self.person
         self.status = Status(**self.status) if self.status else self.status 
+
+    @classmethod
+    def from_json(cls, data):
+        items = {}
+        for k, v in data.items():
+            if k == 'person': 
+                items.update(v)
+            else:
+                items.update({k:v})
+        return cls(**items)
+
+@dataclass(kw_only=True)
+class Coach(Person):
+    """
+    A class to represent a Player.
+
+    Attributes
+    ----------
+    jerseyNumber : str
+        id number of the person
+    job : str
+        job of the coach
+    jobId : str
+        job id of the coach
+    title : str
+        title of the coach
+    parentTeamId : int
+    """
+    jerseyNumber: str
+    job: str
+    jobId: str
+    title: str
+
+    def __post_init__(self):
+        self.position = PrimaryPosition(**self.position) if self.position else self.position
+        self.status = Status(**self.status) if self.status else self.status 
+    
+    @classmethod
+    def from_json(cls, **data):
+        items = {}
+        for k, v in data.items():
+            if k == 'person': 
+                items.update(v)
+            else:
+                items.update({k:v})
+        return cls(**items)
