+++ mlbstatsapi/models/teams/roster.py
@@ -1,6 +1,6 @@
 from typing import List, Union, Optional
 from dataclasses import dataclass, field
-from mlbstatsapi.models.people import Player
+from mlbstatsapi.models.people import Player, Coach
 
 @dataclass
 class Roster:
@@ -25,9 +25,13 @@ class Roster:
     link: str
     teamId: int
     rosterType: str
-    roster: List[Union[Player, dict]] = field(default_factory=dict)
+    roster: List[Union[Player, Coach, dict]] = field(default_factory=dict)
     season: Optional[str] = None
     date: Optional[str] = None
 
     def __post_init__(self):
-        self.roster = [ Player (**player) for player in self.roster ]
+        """Populate roster based on rosterType"""
+        if self.rosterType == "coach":
+            self.roster = [ Coach.from_json(**coach) for coach in self.roster]
+        elif self.rosterType in ['fullRoster', 'active', 'gameday', '40Man', 'fullSeason', 'allTime']:
+            self.roster = [ Player.from_json(**player) for player in self.roster ]
